매장 정보 컴포넌트화
